+++ NEWS.rst
@@ -10,7 +10,8 @@ New features
 Internal changes
 ----------------
 
-* 25193: Improved handling of external configuration files for OS2MO
+* 25193: Improved handling of external configuration files for OS2MO.
+  A warning is no longer triggered on unknown settings.
 
 API changes
 -----------