+++ NEWS.rst
@@ -1,3 +1,43 @@
+Version 0.12.0 (in development)
+===============================
+
+New features
+------------
+
+* 23928: Updated auth to use Flask SAML SSO module.
+  Session is now shared between OS2MO and LoRa.
+
+Internal changes
+----------------
+
+* 25193: Improved handling of external configuration files for OS2MO.
+  A warning is no longer triggered on unknown settings.
+
+API changes
+-----------
+
+
+* TBD
+
+Bug fixes
+---------
+* #24738: Removed sorting and icons for some columns.
+
+
+Known bugs
+----------
+
+
+Version 0.11.1 2018-11-02
+==========================
+
+Bug fixes
+---------
+
+* #25028: Timemachine now shows and updates the organisation unit 
+  view when changing organisation unit
+
+
 Version 0.11.0, 2018-10-30
 ==========================
 
@@ -51,6 +91,10 @@ Bug fixes
   adresses in all fields, and allow reading historical addresses.
 * #24810: Support for Internet Explorer 11.
 * #24570: Sorting now works after performing an update.
+
+
+Known bugs
+----------
 
 
 Version 0.10.1-post1, 2018-10-12